requirements.txt
# Django 핵심
Django==5.1.3
sqlparse==0.5.2
asgiref==3.8.1
tzdata==2024.2

# REST API 개발
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# CORS 처리 (프론트엔드와 통신)
django-cors-headers==4.3.1

# 환경변수 관리 (보안)
python-dotenv==1.0.0

# 데이터베이스
mysqlclient==2.2.0

# Oracle Cloud (프로덕션 환경에서만 필요)
oci==2.119.0

# 배포용 서버
gunicorn==21.2.0
